Add tests for fine-grained locking
One test stresses the locking on a notification object, where two
signalling threads are signalling the same notification object that is
in a waiting state.

The other test stresses the locking on a node's scheduler, where two
signalling threads are signalling different notification objects that
are in waiting states, where the waiting TCBs have the same affinity.